Api: fix mapping user and plant

+++ Api/QuestionRepo/Dto/PlantDto.cs
@@ -4,9 +4,9 @@ namespace QuestionRepo.Dto
 {
     public class PlantDto
     {
-        public int PositionX { get; set; }
-        public int PositionY { get; set; }
-        public int PositionZ { get; set; }
+        public float PositionX { get; set; }
+        public float PositionY { get; set; }
+        public float PositionZ { get; set; }
         public DateTime Datetime { get; set; }
         public int CurrentStage { get; set; }
         public int QuantityHarvested { get; set; }
